+++ Java/src/test/java/com/gildedrose/BackstagePassTest.java
@@ -1,5 +1,6 @@
 package com.gildedrose;
 
+import com.gildedrose.item.Item;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
